GTraits is now an autoload
Automatically declared by the addon into Project Settings
Necessary since GTraits registers traits in its init method. So it must be loaded first! Was not an issue when declaring trait into code since we use the GTraits class, but is now an issue when declaring trait using container through editor...